+++ src/api/resources/index.js
@@ -4,7 +4,7 @@ import config from 'config';
 const up = 'UP';
 const down = 'DOWN';
 
-module.exports = function IndexResource (server) {
+export default function IndexResource (server) {
   server.get('/', (req, res, next) => {
     res.send(200, 'Hi! Like looking at data? Check out https://github.com/madou/armory-react or https://github.com/madou/armory-back if you want to contribute to Guild Wars 2 Armory!');
     return next();
@@ -27,4 +27,4 @@ module.exports = function IndexResource (server) {
 
     return next();
   });
-};
+}